Setting current-context defaults to true
A number of changes in relation to the kube configuration file. The
setting of current-context defaults to true. Also updated the
write-kubeconfig command so that its merges configs as well. Also,
when deleting a cluster it will only try and delete an auto-generated
config file (otherwise it will warn that you need to do this manually).

Issue #29